SOV Slide Generator File Creation/Update
This code lives within a data management file. With the click of a button, it updates and saves a "Share-of-Viewing Slide Generator File" for each market in Comcast Spotlight's West Division. The code opens raw data files for a given market, checks formatting and sourcing, transfers data to a template file, refreshes pivot tables, and saves the file. Included are the main procedures and many (but not all) of the sub-procedures. Assumptions: Nielsen data from Strata software is saved in specific formats and folder structures; a designed Slide-Generator File Template is accessible, with pivot tables and values in expected tabs and cells.