Update SOV Slide Generators Update
This code lives within a "Data Management File" with record-keeping and update buttons. With the click of a button, a "Share-of-Viewing Slide Generator File" is updated for each market in Comcast Spotlight's West Division. The macro opens raw data files for each market, checks formatting and sourcing, transfers data to a Slide Generator File template, makes some adjustments and refreshments, and saves the market-specific files. Assumptions: Nielsen TV ratings data from Strata software is saved in specific formats and folder structures; Slide Generator Template file is accessible, with pivot tables and values in expected tabs and cells. Main procedures and most (but not all) sub-procedures are included.