Batch Run
My team released a "Share-of-Viewing Slide Generator" to the Comcast Spotlight West Division markets. It was later adopted by the the rest of the country. The Excel file has slicers, radio buttons, an aesthetic user interface, and hidden tabs working off of thousands of lines of data- it functions as an app built in Microsoft Excel. The user picks a desire time period, daypart, demographic grouping, and chart settings to see previews of charts showing cable networks' local ratings shares compared to competitor broadcast affiliate channels. The user then clicks on a "generate slide" button to create a PowerPoint slide with the chart and story. We have received very positive feedback and requests for enhancements. One feature that users wanted was the abilitiy to run a "batch" of slides all at once (multiple demos, dayparts, etc.)- This code adds that functionality. The user highlights all desired demos, dayparts, etc. in the slicers and the macro loops through, changing the slicers and and creates a slide for every combinition. The supplied code only includes the userform and "batch" code which I took full ownership of, and references procedures (not included) that build and format the individual PowerPoint slide.