Scarborough Demographic Profiles
I maintain the West Division Network 1-Sheets (a PowerPoint slide with images, descriptions, and localized demographic profiles for each network.) The slides use Scarborough data, which is updated bi-annually. Mannually entering data is tedious, and pasting in data to the PowerPoint tables disrupts the format. This code takes a Scarborough data export of all networks' demographic data for a single market and transfers it to the PowerPoint tables we use. It also prepares the sourcing and makes note if the intab is below our required use value (35.) Assumptions: Scarborough data export in expected format with specific qualitative categories; Empty PowerPoint file with table accessible.